+++ R/dataops_readTable.R
@@ -28,6 +28,7 @@
 #' @param output a "kvConnection" object indicating where the output data should reside.  Must be a \code{\link{localDiskConn}} object if input is a text file on local disk, or a \code{\link{hdfsConn}} object if input is a text file on HDFS.
 #' @param overwrite logical; should existing output location be overwritten? (also can specify \code{overwrite = "backup"} to move the existing output to _bak)
 #' @param params a named list of parameters external to the input data that are needed in \code{postTransFn}
+#' @param packages a vector of R package names that contain functions used in \code{fn} (most should be taken care of automatically such that this is rarely necessary to specify)
 #' @param control parameters specifying how the backend should handle things (most-likely parameters to \code{rhwatch} in RHIPE) - see \code{\link{rhipeControl}} and \code{\link{localDiskControl}}
 #' 
 #' @note For local disk, the file is actually read in sequentially instead of in parallel.  This is because of possible performance issues when trying to read from the same disk in parallel.
@@ -65,6 +66,7 @@ drRead.table <- function(file,
    output = NULL,
    overwrite = FALSE,
    params = NULL,
+   packages = NULL,
    control = NULL,
    ...
    ) {
@@ -126,7 +128,11 @@ drRead.table <- function(file,
    # for now, force factors to be character
    readTabParams$colClasses[readTabParams$colClasses == "factor"] <- "character"
    
-   readTable(file, rowsPerBlock, skip, header, hd, hdText, readTabParams, postTransFn, output, overwrite, params, control)
+   # if the user supplies output as an unevaluated connection
+   # the verbosity can be misleading
+   suppressMessages(output <- output)
+   
+   readTable(file, rowsPerBlock, skip, header, hd, hdText, readTabParams, postTransFn, output, overwrite, params, packages, control)
 }
 
 ############################################################################
@@ -173,7 +179,7 @@ readTable <- function(file, ...)
    UseMethod("readTable", file)
 
 #' @export
-readTable.character <- function(file, rowsPerBlock, skip, header, hd, hdText, readTabParams, postTransFn, output, overwrite, params, control) {
+readTable.character <- function(file, rowsPerBlock, skip, header, hd, hdText, readTabParams, postTransFn, output, overwrite, params, packages, control) {
    
    i <- 1
    for(ff in file) {
@@ -223,7 +229,7 @@ readTable.character <- function(file, rowsPerBlock, skip, header, hd, hdText, re
 }
 
 #' @export
-readTable.hdfsConn <- function(file, rowsPerBlock, skip, header, hd, hdText, readTabParams, postTransFn, output, overwrite, params, control) {
+readTable.hdfsConn <- function(file, rowsPerBlock, skip, header, hd, hdText, readTabParams, postTransFn, output, overwrite, params, packages, control) {
    
    map <- expression({
       tmp <- unlist(map.values)
@@ -245,7 +251,7 @@ readTable.hdfsConn <- function(file, rowsPerBlock, skip, header, hd, hdText, rea
       postTransFn = postTransFn
    )
    
-   ddf(mrExec(ddo(file), map = map, control = control, output = output, overwrite = overwrite, params = c(params, parList)))
+   ddf(mrExec(ddo(file), map = map, control = control, output = output, overwrite = overwrite, params = c(params, parList), packages = packages))
 }
 
 #############################################################################