+++ R/dataops_subset.R
@@ -9,6 +9,7 @@
 #' @param preTransFn a transformation function (if desired) to applied to each subset prior to division - note: this is deprecated - instead use \code{\link{addTransform}} prior to calling divide
 #' @param maxRows the maximum number of rows to return
 #' @param params a named list of parameters external to the input data that are needed in the distributed computing (most should be taken care of automatically such that this is rarely necessary to specify)
+#' @param packages a vector of R package names that contain functions used in \code{fn} (most should be taken care of automatically such that this is rarely necessary to specify)
 #' @param control parameters specifying how the backend should handle things (most-likely parameters to \code{rhwatch} in RHIPE) - see \code{\link{rhipeControl}} and \code{\link{localDiskControl}}
 #' @param verbose logical - print messages about what is being done
 #' 
@@ -23,17 +24,11 @@ drSubset <- function(data,
    preTransFn = NULL,
    maxRows = 500000,
    params = NULL,
+   packages = NULL,
    control = NULL,
    verbose = TRUE
 ) {
-   # data <- divide(iris, by = "Species")
-   # subset <- expression(Sepal.Length > 5)
-   # select <- NULL
-   # drop <- FALSE
-   # preTransFn <- flatten
-   # maxRows <- 500000
-   # params <- NULL; control <- NULL; verbose <- TRUE
-
+   
    if(!inherits(data, "ddf")) {
       if(verbose)
          message("* Input data is not 'ddf' - attempting to cast it as such")
@@ -76,18 +71,18 @@ drSubset <- function(data,
       drop = drop
    )
    
+   parList <- c(parList, list(
+      applyTransform = applyTransform,
+      setupTransformEnv = setupTransformEnv, 
+      kvApply = kvApply
+   ))
+   
    if(! "package:datadr" %in% search()) {
       if(verbose)
          message("* ---- running dev version - sending datadr functions to mr job")
-      # parList <- c(parList, list())
-      
-      setup <- expression({
-         suppressWarnings(suppressMessages(library(data.table)))
-      })
+      packages <- c(packages, "data.table")
    } else {
-      setup <- expression({
-         suppressWarnings(suppressMessages(library(datadr)))
-      })
+      packages <- c(packages, "datadr", "data.table")
    }
    
    map <- expression({
@@ -124,10 +119,10 @@ drSubset <- function(data,
    )
    
    res <- mrExec(data,
-      setup     = setup,
       map       = map, 
       reduce    = reduce, 
       params    = c(params, parList),
+      packages  = packages,
       control   = control
    )