+++ R/mapreduce_kvLocalDisk.R
@@ -69,7 +69,7 @@ mrExecInternal.kvLocalDiskList <- function(data, setup = NULL, map = NULL, reduc
    
    mapFn <- function(fl) {
       mapEnv <- new.env() # parent = baseenv())
-      assign(".mr_packages", params$.mr_packages, mapEnv)
+      assign("mr___packages", params$mr___packages, mapEnv)
       eval(setup, envir = mapEnv)
       
       # add a collect, counter functions to the environment
@@ -224,7 +224,7 @@ mrExecInternal.kvLocalDiskList <- function(data, setup = NULL, map = NULL, reduc
    # if output is a character string, construct a localDiskConn connection from it
    if(!inherits(output, "localDiskConn")) {
       if(is.character(output)) {
-         output <- localDiskConn(output, nBins = floor(length(outputKeyHash) / 1000))
+         output <- localDiskConn(output, nBins = floor(length(outputKeyHash) / 1000), verbose = FALSE)
       } else {
          output <- localDiskConn(tempfile(paste("job", jobNum, "_", sep = "")), nBins = floor(length(outputKeyHash) / 1000), autoYes = TRUE, verbose = FALSE)
       }
@@ -312,17 +312,22 @@ localDiskControl <- function(
    reduce_buff_size_bytes = 10485760,
    map_temp_buff_size_bytes = 10485760
 ) {
-   list(
+   structure(list(
       cluster = cluster,
       map_buff_size_bytes = map_buff_size_bytes,
       reduce_buff_size_bytes = reduce_buff_size_bytes,
       map_temp_buff_size_bytes = map_temp_buff_size_bytes
-   )
+   ), class = c("localDiskControl", "list"))
 }
 
 #' @export
 defaultControl.kvLocalDisk <- function(x) {
-   localDiskControl()
+   res <- getOption("defaultLocalDiskControl")
+   if(inherits(res, "localDiskControl")) {
+      return(res)
+   } else {
+      return(localDiskControl())
+   }
 }
 
 # this will be used for both map and reduce